initial cut of clear ones
